// src/configs/config.enum.ts
/* eslint-disable no-unused-vars */
export enum CONFIG {
  DOMAIN_URL = "DOMAIN_URL",
  ACCESS_CONTROLL_ADDRESS = "ACCESS_CONTROLL_ADDRESS",
  ACTIVE_NFT_ADDRESS = "ACTIVE_NFT_ADDRESS",
  MANAGER_POOL_ADDRESS = "MANAGER_POOL_ADDRESS",
  MARKETPLACE_ADDRESS = "MARKETPLACE_ADDRESS",
  UIT_TOKEN_ADDRESS = "UIT_TOKEN_ADDRESS",
  UIT_NFT_TOKEN_ADDRESS = "UIT_NFT_TOKEN_ADDRESS",
}